Revamp UI/layout
The main part of this is that it creates a top-level, global, shared
sidenav component. Previously, the sidenav was entangled with the map
component.

This also improves/fixes some top-level layout/positioning issues.

This is different take on the sidenav aspect of PR #2.

First pass; will probably require additional thought and tweaking.