Name layers in GeoServer after Django content type names
That is, instead of naming the building layer "buildings", it's now
named "buildings.building". This has two purposes:

1. Removes the possibility of layer name conflicts (although this isn't
   really a big concern right now)
2. Makes finding the database record corresponding to a GeoServer
   feature easy because Geoserver generates feature IDs as
   "{layer name}.{pk}" (this is the main reason for this change)

For example, for a GeoServer feature with the ID "buildings.building.1",
we can easily look up the corresponding DB record since we know exactly
which Django app & model that ID refers to.