Add a search-by-ID endpoint
The URL pattern is `/search/{app}.{model}.{pk}`.

This allows retrieval of a specific item in the same format used by the
main search endpoint. The intent is to enable simpler, more-generic
front end search code.

This also changes the URL of the main search endpoint so that it ends
with a slash.