Simplify click-to-show-feature-info event handling
Instead of registering a separate map click event handler, register a
handler for the select interaction's `select` event. This is simpler and
more-closely matches the intent of "show info when selected".

This also fixes an issue where the selected feature would be unselected
when the map was auto-panned (i.e., because it was too close to a map
edge). That would happen because the map click handler would run first:
the feature would be panned and the select interaction wouldn't see it.