Move map base layer switcher to bottom left of map
This adds top right, bottom right, and bottom left map control regions.
There's no top left region because the search box is in that corner.

Moving the base layer switcher to the bottom left opens up the top right
for a feature layer selector.

We'll probably want to change the styling of the base layer switcher to
make it more prominent. And when the sidenav is open, the layer switcher
should be moved to the right so it's not hidden.